rename issue{61,62}
