--- tests/fixedbugs/issue61/issue61.s
@@ -1,5 +1,0 @@
-// Code generated by command: go run asm.go -out issue61.s. DO NOT EDIT.
-
-// func private()
-TEXT ·private(SB), $0
-	RET
