--- tests/fixedbugs/issue61/stub.go
@@ -1,5 +1,0 @@
-// Package issue61 tests for using Implement() with an unexported function.
-package issue61
-
-// private serves as an example unexported function to implement.
-func private()
