+++ tests/fixedbugs/issue62/issue62.s
@@ -1,0 +1,5 @@
+// Code generated by command: go run asm.go -out issue62.s. DO NOT EDIT.
+
+// func private()
+TEXT ·private(SB), $0
+	RET
