+++ tests/fixedbugs/issue62/issue62_test.go
@@ -1,10 +1,10 @@
-package issue61
+package issue62
 
 import (
 	"testing"
 )
 
-//go:generate go run asm.go -out issue61.s
+//go:generate go run asm.go -out issue62.s
 
 func TestPrivate(t *testing.T) {
 	private()
